requirements.txt
chardet==5.2.0
openpyxl==3.1.5
pillow==12.0.0
pypdf==6.5.0
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
python-docx==1.2.0
python-pptx==1.0.2
